backend/pyproject.toml
[project]
name = "uncloud-medical-rag-backend"
version = "0.1.0"
description = "Medical-grade provenance-first GraphRAG backend scaffold for the IKNL hackathon"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.116.0",
  "httpx>=0.28.0",
  "orjson>=3.10.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.6.0",
  "pyyaml>=6.0.2",
  "uvicorn>=0.35.0",
]

[project.optional-dependencies]
anthropic = [
  "anthropic>=0.72.0",
]
docs = [
  "docling>=2.70.0",
]
firecrawl = [
  "firecrawl-py>=4.0.0",
]
graph = [
  "neo4j>=5.28.0",
  "neo4j-graphrag>=1.10.0",
]
dev = [
  "pytest>=8.4.0",
]

[tool.uv]
package = true

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
